requirements.txt
# PhoBERT Contract Processing Requirements
# Yêu cầu thư viện cho dự án xử lý hợp đồng với PhoBERT

# Core Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Tokenization (for Vietnamese)
vncorenlp>=1.0.3
pyvi>=0.1.1

# Training & Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Logging & Monitoring (optional)
tensorboard>=2.13.0
wandb>=0.15.0

# Metrics
seqeval>=1.2.2  # For NER evaluation
rouge-score>=0.1.2  # For summarization evaluation
datasets>=2.14.0  # For data processing and metrics

# Testing
pytest>=7.4.0

# Utilities
python-dotenv>=1.0.0

# Optional: For better performance
accelerate>=0.20.0  # For distributed training
# deepspeed>=0.9.0    # For optimization

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0  # For HTML templates

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebooks
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

# Model Export & Production
onnx>=1.14.0
onnxruntime>=1.15.0

# Efficient Fine-tuning (optional)
peft>=0.5.0  # For LoRA fine-tuning
